requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
numpy==1.16.3
pandas==0.24.2
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.22.0
six==1.12.0
urllib3==1.25.2
xlrd==1.2.0
